chat-back/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
python-dotenv>=1.0.0
httpx>=0.25.0
langchain>=0.1.0
langchain-core>=0.1.0
langgraph>=0.0.20
langsmith>=0.0.60
starlette>=0.27.0
typing-extensions>=4.8.0
loguru>=0.7.0
